move functions.dart
