+++ packages/flame_jenny/jenny/lib/src/structure/expressions/functions/functions.dart
